+++ packages/typescript/src/typescript.ts
@@ -77,6 +77,11 @@ export function typescript(configFilePath: string, extraCompilerOptions: ts.Comp
   });
 }
 
+// TypeScript throws a 6307 error when it need to access type information from a file
+// that wasn't included by the tsconfig. This happens whenever we run the compiler on
+// a subset of files, so we need to filter out those errors!
+const CODE_FILE_NOT_INCLUDED = 6307;
+
 /** @internal Definitely not stable! Please don't use! */
 export function typescriptΔ(configFilePath: string, extraCompilerOptions: ts.CompilerOptions = {}): BettererFileTest {
   if (!configFilePath) {
@@ -102,8 +107,11 @@ export function typescriptΔ(configFilePath: string, extraCompilerOptions: ts.Co
       ...extraCompilerOptions
     };
     config.compilerOptions = fullCompilerOptions;
-    config.files = filePaths;
-    delete config.include;
+
+    if (!config.compilerOptions.incremental) {
+      config.files = filePaths;
+      delete config.include;
+    }
 
     const compilerHost = ts.createCompilerHost(fullCompilerOptions);
     const configHost: ts.ParseConfigHost = {
@@ -135,13 +143,16 @@ export function typescriptΔ(configFilePath: string, extraCompilerOptions: ts.Co
     ]);
 
     allDiagnostics
-      .filter(({ file, start, length }) => file && start != null && length != null)
-      .forEach((diagnostic) => {
-        const { start, length } = diagnostic as ts.DiagnosticWithLocation;
-        const source = (diagnostic as ts.DiagnosticWithLocation).file;
-        const { fileName } = source;
-        const file = fileTestResult.addFile(fileName, source.getFullText());
-        const message = ts.flattenDiagnosticMessageText(diagnostic.messageText, NEW_LINE);
+      .filter((diagnostic): diagnostic is ts.DiagnosticWithLocation => {
+        const { file, start, length } = diagnostic;
+        return file != null && start != null && length != null;
+      })
+      .filter(({ file, code }) => {
+        return filePaths.includes(file.fileName) && code !== CODE_FILE_NOT_INCLUDED;
+      })
+      .forEach(({ start, length, file: source, messageText }) => {
+        const file = fileTestResult.addFile(source.fileName, source.getFullText());
+        const message = ts.flattenDiagnosticMessageText(messageText, NEW_LINE);
         file.addIssue(start, start + length, message);
       });
   });