+++ packages/betterer/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.4.0](https://github.com/phenomnomnominal/betterer/compare/v4.3.8...v4.4.0) (2021-06-18)
+
+
+### Bug Fixes
+
+* **betterer 🐛:** make sure results file stays in the same order ([#766](https://github.com/phenomnomnominal/betterer/issues/766)) ([ad752d0](https://github.com/phenomnomnominal/betterer/commit/ad752d09701ca7d9855f22e0ce72d201d5f09915))
+
+
+### Features
+
+* **betterer ✨:** add precommit mode ([#772](https://github.com/phenomnomnominal/betterer/issues/772)) ([da59df3](https://github.com/phenomnomnominal/betterer/commit/da59df310136699ee213895936028e43e4a54abc))
+
+
+
+
+
+## [4.3.8](https://github.com/phenomnomnominal/betterer/compare/v4.3.7...v4.3.8) (2021-06-16)
+
+
+### Bug Fixes
+
+* **betterer 🐛:** even more cache fixes ([#746](https://github.com/phenomnomnominal/betterer/issues/746)) ([ee4f944](https://github.com/phenomnomnominal/betterer/commit/ee4f9446046939ae20892daed8f51530f76c2f9a))
+
+
+
+
+
 ## [4.3.6](https://github.com/phenomnomnominal/betterer/compare/v4.3.5...v4.3.6) (2021-06-11)